fix: migrate Storybook main config from CommonJS to ESM
